+++ R/to_text.R
@@ -21,3 +21,51 @@ paste_conjunction <- function(x, conjunction = "and") {
   }
   return(c)
 }
+
+#' Spell out numbers
+#'
+#' Spell out and concatenate a numeric vector if any of the numbers are >10.
+#' Vectors of length >1 are concatenated with an optional conjunction.
+#' `spell_numbers()` wraps `xfun::numbers_to_words()` and adds some additional logic.
+#'
+#' @param x Numeric vector to be converted to text
+#' @param conjunction Conjunction to be used behind the oxford comma. Usually a character vector vector of length one, but can be any object that can be converted to a character vector
+#' @returns Character vector of length 1
+#' @export
+#' @examples
+#' spell_numbers(2)
+#' spell_numbers(20)
+#' spell_numbers(c(1, 2, 3))
+#' spell_numbers(c(1, 2, 30))
+#' spell_numbers(c(1, 2, 3), conjunction = "and")
+#' spell_numbers(c(1, 2, 30), conjunction = "and")
+spell_numbers <- function(x, conjunction = NULL) {
+  if (!is.numeric(x)) {
+    cli::cli_abort(
+      c(
+        "!" = "{.arg x} must be numeric.",
+        "x" = "You supplied {.type {x}}"
+      )
+    )
+  }
+
+  if (length(x) == 1) {
+    x <- ifelse(
+      x < 10,
+      xfun::numbers_to_words(x, cap = FALSE, hyphen = TRUE, and = FALSE),
+      as.character(x)
+    )
+  } else {
+    if (all(x < 10)) {
+      x <- purrr::map(x, spell_numbers) |>
+        stringi::stri_join_list()
+    }
+    if (!is.null(conjunction)) {
+      x <- paste_conjunction(x = x, conjunction = conjunction)
+    } else {
+      x <- stringi::stri_join(x, collapse = ", ")
+    }
+  }
+
+  return(x)
+}